Add model definitions for event and ticket type
Build out definitions for the event and ticket type models and attach
them to the Express request object in middleware for easy access in the
routes.

An event is defined as any sort of performance or activity requiring
ticket sales, for example, a concert.

A ticket type is defined as a block and price level of tickets to be
sold for an event.

For example:

event -> ODESZA Concert @ Commodore Ballroom, December 02
ticket type -> 200 VIP Tickets @ $100 each
ticket type -> 1000 General Admission tickets @ $50 each